re-organize server
