+++ app.js
@@ -1,18 +1,8 @@
-const express = require('express');
-const app = express();
 const { db, seed, Entry } = require('./db/index');
-const port = process.env.PORT || 3000
 const path = require('path')
-const bodyParser = require('body-parser');
-const socketio = require('socket.io');
-
+const app = require('./server');
 const Filter = require('bad-words');   // this is my bad word filter
-
 const badFilter = new Filter();
-
-
-app.use(express.static('public'));
-app.use(bodyParser.json());
 
 app.get('/api/entry', (req, res, next) => {
   Entry.findAll({
@@ -31,13 +21,9 @@ app.get('/', (req, res, next) => {
 
 app.post('/api/user/entry', (req, res, next) => {
   Entry.create(req.body)
-  .then(res.sendStatus(200))
-  .catch(next);
+    .then(res.sendStatus(200))
+    .catch(next);
 })
-
-const server = app.listen(port, () => {
-  console.log(`I am listening on port, ${port}`);
-});
 
 const init = async () => {
   await db.sync({ force: true })
@@ -46,19 +32,6 @@ const init = async () => {
 
 init();
 
-//========== Web Socket Logic ==================================== aw yea ===============\\
-
-const io = socketio(server);
-
-io.on('connect', socket => {
-  console.log(socket.id, 'is connected')
-
-  //the server listens for a new entry submission and then broadcasts that new entry obj (before actually entering it in the db) to other connected browsers
-  socket.on('entry', entry => {
-    socket.broadcast.emit('entry', entry);
-  })
-})
 
 
 
-module.exports = app;
