+++ package.json
@@ -10,13 +10,16 @@
     "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
+    "bad-words": "^1.6.3",
     "body-parser": "^1.18.3",
     "express": "^4.16.3",
     "nodemon": "^1.18.3",
     "pg": "^7.4.3",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
+    "react-router-dom": "^4.3.1",
     "sequelize": "^4.38.0",
+    "socket.io": "^2.1.1",
     "webpack": "^4.16.5",
     "webpack-cli": "^3.1.0"
   },
